rag-backend/requirements.txt
fastapi
uvicorn
faiss-cpu
transformers
torch
pypdf
python-multipart
